backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
faster-whisper==1.0.3
numpy==2.1.2
soundfile==0.12.1
requests==2.32.3
openai==1.51.2
azure-cognitiveservices-speech==1.38.0
